Title: 선분 위의 점, Time: 181ms, Memory: 13MB, Status: Wrong Answer - CodeTree

+++ 240306/선분 위의 점/point-on-the-line-segment.java
@@ -2,7 +2,7 @@ import java.io.*;
 import java.util.*;
 
 public class Main {
-    static int[] arr;
+    static Long[] arr;
     static int n;
     static StringBuilder sb = new StringBuilder();
 
@@ -14,10 +14,10 @@ public class Main {
         n = Integer.parseInt(st.nextToken());
         int m = Integer.parseInt(st.nextToken());
 
-        arr = new int[n + 1];
+        arr = new Long[n + 1];
         st = new StringTokenizer(br.readLine());
         for(int i = 1; i <= n; i++){
-            arr[i] = Integer.parseInt(st.nextToken());
+            arr[i] = Long.parseLong(st.nextToken());
         }
 
         for(int i = 0; i < m; i++){
@@ -38,13 +38,13 @@ public class Main {
         int minIdx = n+1;
 
         while(left <= right){
-            Long  mid = (left + right) / 2;
+            long mid = (long) ((left + right) / 2);
 
-            if(arr[mid] >= target){
-                right = mid - 1;
-                minIdx = Math.min(minIdx, mid);
+            if(arr[(int) mid] >= target){
+                right = (int) (mid - 1);
+                minIdx = (int) Math.min(minIdx, mid);
             }else{
-                left = mid + 1;
+                left = (int) (mid + 1);
             }
         }
         return minIdx;
@@ -56,13 +56,13 @@ public class Main {
         int minIdx = n+1;
 
         while(left <= right){
-            int mid = (left + right) / 2;
-            if(arr[mid] > target){
-                right = mid - 1;
-                minIdx = Math.min(minIdx, mid);
+            long mid = (long) ((left + right) / 2);            
+            if(arr[(int)mid] > target){
+                right = (int) (mid - 1);
+                minIdx = (int) Math.min(minIdx, mid);
             }
             else{
-                left = mid + 1;
+                left = (int) (mid + 1);
             }
 
         }
